+++ packages/react-native-renderer/src/ReactFabric.js
@@ -28,6 +28,7 @@ import {
 import {createPortal} from 'shared/ReactPortal';
 import {setBatchingImplementation} from 'events/ReactGenericBatching';
 import ReactVersion from 'shared/ReactVersion';
+import createEventComponent from 'shared/createEventComponent';
 
 import NativeMethodsMixin from './NativeMethodsMixin';
 import ReactNativeComponent from './ReactNativeComponent';
@@ -38,6 +39,8 @@ import {LegacyRoot} from 'shared/ReactRootTags';
 import ReactSharedInternals from 'shared/ReactSharedInternals';
 import getComponentName from 'shared/getComponentName';
 import warningWithoutStack from 'shared/warningWithoutStack';
+
+import {enableEventAPI} from 'shared/ReactFeatureFlags';
 
 const ReactCurrentOwner = ReactSharedInternals.ReactCurrentOwner;
 
@@ -152,7 +155,12 @@ const ReactFabric: ReactFabricType = {
     // Used as a mixin in many createClass-based components
     NativeMethodsMixin: NativeMethodsMixin(findNodeHandle, findHostInstance),
   },
+  unstable_createEventComponent: undefined,
 };
+
+if (enableEventAPI) {
+  ReactFabric.unstable_createEventComponent = createEventComponent;
+}
 
 injectIntoDevTools({
   findFiberByHostInstance: getClosestInstanceFromNode,
